feat: Export `ValidationMessage` type (#346)
