+++ src/components/TextInput/index.tsx
@@ -1,1 +1,4 @@
+import { ValidationMessage as ValidationMessageType } from './styled';
+
 export { Component as TextInput } from './component';
+export type ValidationMessage = ValidationMessageType;
